+++ src/accountsservice/usersmodel.h
@@ -76,7 +76,7 @@ public:
 private:
     Q_DECLARE_PRIVATE(UsersModel)
     Q_PRIVATE_SLOT(d_func(), void _q_userAdded(UserAccount *account))
-    Q_PRIVATE_SLOT(d_func(), void _q_userDeleted(quint32 uid))
+    Q_PRIVATE_SLOT(d_func(), void _q_userDeleted(qlonglong uid))
 };
 
 }
